requirements.txt
pillow==7.2.0
tqdm>=4.36.1
pandas>=0.24.1
numpy>=1.16.5
scikit-image==0.17.2
opencv==4.1.1
future~=0.18.2
opencv-python~=4.3.0.36
jinja2~=2.11.2